renaming pages2->pages
