+++ Multiplex_Analysis_Web_Apps.py
@@ -15,31 +15,31 @@ import streamlit_utils
 import platform_io
 
 # Note if any of the following imports having "  # slow" are not commented out, there is a delay in running the forking test
-from pages2 import data_import_and_export
-from pages2 import datafile_format_unifier
-from pages2 import open_file
-from pages2 import feature_creation
-from pages2 import robust_scatter_plotter
-from pages2 import multiaxial_gating
-from pages2 import thresholded_phenotyping  # slow due to things ultimately importing umap
-from pages2 import adaptive_phenotyping
-from pages2 import Pheno_Cluster_a  # "slow" for forking test initialization
-from pages2 import Pheno_Cluster_b  # "slow" for forking test initialization
-from pages2 import Tool_parameter_selection
-from pages2 import Run_workflow
-from pages2 import Display_individual_ROI_heatmaps
-from pages2 import Display_average_heatmaps
-from pages2 import Display_average_heatmaps_per_annotation
-from pages2 import Display_ROI_P_values_overlaid_on_slides
-from pages2 import Neighborhood_Profiles  # slow due to things ultimately importing umap
-from pages2 import UMAP_Analyzer  # slow due to things ultimately importing umap
-from pages2 import Clusters_Analyzer  # slow due to things ultimately importing umap
-from pages2 import memory_analyzer
-from pages2 import radial_bins_plots
-from pages2 import radial_profiles_analysis
-from pages2 import preprocessing
-from pages2 import results_transfer
-# from pages2 import forking_test
+from pages import data_import_and_export
+from pages import datafile_format_unifier
+from pages import open_file
+from pages import feature_creation
+from pages import robust_scatter_plotter
+from pages import multiaxial_gating
+from pages import thresholded_phenotyping  # slow due to things ultimately importing umap
+from pages import adaptive_phenotyping
+from pages import Pheno_Cluster_a  # "slow" for forking test initialization
+from pages import Pheno_Cluster_b  # "slow" for forking test initialization
+from pages import Tool_parameter_selection
+from pages import Run_workflow
+from pages import Display_individual_ROI_heatmaps
+from pages import Display_average_heatmaps
+from pages import Display_average_heatmaps_per_annotation
+from pages import Display_ROI_P_values_overlaid_on_slides
+from pages import Neighborhood_Profiles  # slow due to things ultimately importing umap
+from pages import UMAP_Analyzer  # slow due to things ultimately importing umap
+from pages import Clusters_Analyzer  # slow due to things ultimately importing umap
+from pages import memory_analyzer
+from pages import radial_bins_plots
+from pages import radial_profiles_analysis
+from pages import preprocessing
+from pages import results_transfer
+# from pages import forking_test
 
 # Configure logging
 logging.basicConfig(level=logging.INFO,
