+++ pages/Display_average_heatmaps.py
