+++ pages/Display_average_heatmaps_per_annotation.py
