+++ pages/adaptive_phenotyping.py
@@ -3,7 +3,7 @@ import streamlit as st
 import numpy as np
 import pandas as pd
 import plotly.graph_objects as go
-from pages2 import multiaxial_gating
+from pages import multiaxial_gating
 import utils
 
 # Global variable
