+++ pages/memory_analyzer.py
