+++ pages/skeleton2.py
